ecommerce/requirements.txt
kafka-python==2.0.2
faker==19.3.0
jsonpickle==3.0.2
boto3==1.28.0
pyspark==3.5.0
streamlit==1.28.0
plotly==5.18.0
pandas==2.0.3
